+++ carveracontroller/main.py
@@ -2629,6 +2629,9 @@ class Makera(RelativeLayout):
     light_state = LightProperty(False)
 
     played_lines = 0
+    _remaining_anchor_sec = 0.0
+    _remaining_anchor_time = 0.0
+    _progress_smooth_clock = None
 
     show_update = True
     fw_upd_text = ''
@@ -2786,9 +2789,11 @@ class Makera(RelativeLayout):
 
         # init gcode viewer
         self.gcode_viewer = GCodeViewer()
+        self.gcode_viewer.high_precision_time_estimate = Config.getboolean('carvera', 'high_precision_reamining_time_estimate', fallback=True)
         self.gcode_viewer_container.add_widget(self.gcode_viewer)
         self.gcode_viewer.set_frame_callback(self.gcode_play_call_back)
         self.gcode_viewer.set_play_over_callback(self.gcode_play_over_call_back)
+        self.gcode_viewer.time_estimate_progress_callback = self._on_time_estimate_progress
 
         # init settings
         self.config = ConfigParser()
@@ -4644,6 +4649,28 @@ class Makera(RelativeLayout):
         self.progress_popup.dismiss()
 
     # --------------------------------------------------------------`---------
+    def _on_time_estimate_progress(self, state, percent):
+        """Callback for GcodeViewer time estimate computation: show progress popup while parsing feed speeds."""
+        if state == 'start':
+            self.progressStart(tr._('Calculating run time time estimate...'), None)
+        elif state == 'progress':
+            self.progressUpdate(percent, '', True)
+        elif state == 'done':
+            self.progressFinish()
+
+    # --------------------------------------------------------------`---------
+    def _update_progress_smooth(self, dt):
+        """Refresh elapsed/remaining display every second while playing."""
+        app = App.get_running_app()
+        if not app.playing or (not app.selected_remote_filename and not app.selected_local_filename) or not self.selected_file_line_count:
+            return
+        remaining_display = max(0.0, self._remaining_anchor_sec - (time.time() - self._remaining_anchor_time))
+        filename = os.path.basename(app.selected_remote_filename or app.selected_local_filename)
+        self.progress_info = ' {} ( {}/{} - {}%, {} elapsed, {} to go )'.format(
+            filename, self.played_lines, self.selected_file_line_count, int(self.wpb_play.value),
+            Utils.second2hour(CNC.vars["playedseconds"]), Utils.second2hour(int(remaining_display)))
+
+    # --------------------------------------------------------------`---------
     def updateCompressProgress(self, value):
         Clock.schedule_once(partial(self.progressUpdate, value * 100.0 / self.fileCompressionBlocks, '', True), 0)
         if value == self.fileCompressionBlocks:
@@ -4973,6 +5000,10 @@ class Makera(RelativeLayout):
                 self.wpb_leveling.value = 0
                 self.wpb_play.value = 0
                 self.progress_info = ""
+                # Stop smooth progress updates
+                if self._progress_smooth_clock is not None:
+                    self._progress_smooth_clock.cancel()
+                    self._progress_smooth_clock = None
 
                 last_job_elapsed = ""
                 if CNC.vars["playedseconds"] > 0:
@@ -4986,23 +5017,26 @@ class Makera(RelativeLayout):
                     self.progress_info = tr._(' No Remote File Selected') + last_job_elapsed
             else:
                 app.playing = True
-                # playing file remotely
                 if self.played_lines != CNC.vars["playedlines"]:
                     self.played_lines = CNC.vars["playedlines"]
                     self.wpb_play.value = CNC.vars["playedpercent"]
-                    self.progress_info = ''
                     if (app.selected_remote_filename != '' or app.selected_local_filename != '') and self.selected_file_line_count > 0:
-                        # update gcode list
                         self.gcode_rv.set_selected_line(self.played_lines)
-                        # update gcode viewer
                         self.gcode_viewer.set_distance_by_lineidx(self.played_lines, 0.5)
-                        # update progress info
-                        self.progress_info = os.path.basename(app.selected_remote_filename if app.selected_remote_filename != '' else app.selected_local_filename) + ' ( {}/{} - {}%, {} elapsed'.format( \
-                                                     self.played_lines, self.selected_file_line_count, int(self.wpb_play.value), Utils.second2hour(CNC.vars["playedseconds"]))
-                        if self.wpb_play.value > 0:
-                            self.progress_info = self.progress_info + ', {} to go )'.format(Utils.second2hour((100 - self.wpb_play.value) * CNC.vars["playedseconds"] / self.wpb_play.value))
+                        remaining_sec = self.gcode_viewer.get_remaining_time_by_lineidx(self.played_lines, 0.5)
+                        if remaining_sec is not None and remaining_sec >= 0:
+                            ov_feed = CNC.vars.get("OvFeed", 100) or 100
+                            self._remaining_anchor_sec = remaining_sec / (ov_feed / 100.0)
+                        elif self.wpb_play.value > 0:
+                            self._remaining_anchor_sec = (100 - self.wpb_play.value) * CNC.vars["playedseconds"] / self.wpb_play.value
+                        elif self.gcode_viewer.total_time > 0:
+                            ov_feed = CNC.vars.get("OvFeed", 100) or 100
+                            self._remaining_anchor_sec = self.gcode_viewer.total_time / (ov_feed / 100.0)
                         else:
-                            self.progress_info = self.progress_info + ' )'
+                            self._remaining_anchor_sec = 0.0
+                        self._remaining_anchor_time = now
+                if (app.selected_remote_filename != '' or app.selected_local_filename != '') and self.selected_file_line_count > 0 and self._progress_smooth_clock is None:
+                    self._progress_smooth_clock = Clock.schedule_interval(self._update_progress_smooth, 1.0)
                 # playing margin
                 if CNC.vars["atc_state"] == 4:
                     self.wpb_margin.value += 14
@@ -5660,6 +5694,9 @@ class Makera(RelativeLayout):
         if "log_sent_receive" in self.controller_setting_change_list:
             self.controller.log_sent_receive = self.controller_setting_change_list.get("log_sent_receive")
 
+        if "high_precision_reamining_time_estimate" in self.controller_setting_change_list:
+            self.gcode_viewer.high_precision_time_estimate = self.controller_setting_change_list.get("high_precision_reamining_time_estimate")
+
 
     # -----------------------------------------------------------------------
     def open_setting_restore_confirm_popup(self):
@@ -6168,6 +6205,7 @@ def set_config_defaults(default_lang):
     if not Config.has_option('carvera', 'invert_y_axis_jogging'): Config.set('carvera', 'invert_y_axis_jogging', '0')
     if not Config.has_option('carvera', 'use_higher_baud'): Config.set('carvera', 'use_higher_baud', '0')
     if not Config.has_option('carvera', 'usb_baud_rate'): Config.set('carvera', 'usb_baud_rate', '1500000')
+    if not Config.has_option('carvera', 'high_precision_reamining_time_estimate'): Config.set('carvera', 'high_precision_reamining_time_estimate', '1')
     if not Config.has_option('carvera', 'background_image'): Config.set('carvera', 'background_image', 'None')
     if not Config.has_option('graphics', 'allow_screensaver'): Config.set('graphics', 'allow_screensaver', '0')
     if not Config.has_option('graphics', 'height'): Config.set('graphics', 'height', '1440')